fixed zip code and phone number tables

+++ migrations/20181120182850-create-users.js
@@ -17,11 +17,11 @@ module.exports = {
        name_first:{
          type: Sequelize.STRING,
        },
-             name_last: {
-               type: Sequelize.STRING,
+         name_last: {
+          type: Sequelize.STRING,
              },
       phone: {
-        type: Sequelize.INTEGER
+        type: Sequelize.NUMERIC
       },
       email: {
         type: Sequelize.STRING
